requirements.txt
fastapi
uvicorn
feedparser
listparser
pytest
httpx
fastapi-mcp
pyyaml
pytest-mock
jsonschema
cachetools
respx
